# backend/nestjs/generated-express/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: dev@avis.it
    name: Team Tecnico AVIS
    url: https://www.avis.it
  description: |
    API per l'app mobile AVIS (Italia) per la gestione delle donazioni di sangue.
    Lingua primaria: italiano (supporto inglese opzionalmente). Implementazione MVP
    ispirata al NHS Give Blood (booking, centri, storico donazioni, notifiche).
  title: AVIS Donazioni API
  version: 1.0.0
servers:
- description: API produzione (esempio)
  url: https://api.avis.it/v1
- description: Ambiente di staging
  url: https://staging.api.avis.it/v1
tags:
- description: Autenticazione e gestione account donatore
  name: auth
- description: Profili donatori e storico
  name: donors
- description: Centri di raccolta e disponibilità
  name: centers
- description: Prenotazioni
  name: appointments
- description: Risultati esami
  name: tests
- description: Notifiche push e subscription
  name: notifications
paths:
  /auth/register:
    post:
      operationId: authRegisterPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
          description: Donatore creato
        "400":
          description: Bad request
      summary: Registra un nuovo donatore
      tags:
      - auth
      x-eov-operation-handler: controllers/AuthController
  /auth/login:
    post:
      operationId: authLoginPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_auth_login_post_200_response"
          description: Token di accesso
      summary: Login (email + password) / SSO
      tags:
      - auth
      x-eov-operation-handler: controllers/AuthController
  /donors/{donorId}:
    get:
      operationId: donorsDonorIdGET
      parameters:
      - explode: false
        in: path
        name: donorId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
          description: Profilo donatore
      security:
      - bearerAuth: []
      summary: Recupera profilo donatore
      tags:
      - donors
      x-eov-operation-handler: controllers/DonorsController
    put:
      operationId: donorsDonorIdPUT
      parameters:
      - explode: false
        in: path
        name: donorId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorUpdate"
        required: true
      responses:
        "200":
          description: Donatore aggiornato
      security:
      - bearerAuth: []
      summary: Aggiorna dati donatore
      tags:
      - donors
      x-eov-operation-handler: controllers/DonorsController
  /centers:
    get:
      operationId: centersGET
      parameters:
      - explode: true
        in: query
        name: lat
        required: false
        schema:
          format: double
          type: number
        style: form
      - explode: true
        in: query
        name: lng
        required: false
        schema:
          format: double
          type: number
        style: form
      - explode: true
        in: query
        name: radius_km
        required: false
        schema:
          default: 50
          type: integer
        style: form
      - explode: true
        in: query
        name: region
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Center"
                type: array
          description: Lista centri
      summary: Lista centri (filtro per regione/coordinate/tipo)
      tags:
      - centers
      x-eov-operation-handler: controllers/CentersController
  /centers/{centerId}/availability:
    get:
      operationId: centersCenterIdAvailabilityGET
      parameters:
      - explode: false
        in: path
        name: centerId
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: date_from
        required: false
        schema:
          format: date
          type: string
        style: form
      - explode: true
        in: query
        name: date_to
        required: false
        schema:
          format: date
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Slot"
                type: array
          description: Slots disponibili
      summary: Disponibilità (slots) per centro
      tags:
      - centers
      x-eov-operation-handler: controllers/CentersController
  /appointments:
    get:
      operationId: appointmentsGET
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Appointment"
                type: array
          description: Lista appuntamenti
      security:
      - bearerAuth: []
      summary: Lista appuntamenti del donatore
      tags:
      - appointments
      x-eov-operation-handler: controllers/AppointmentsController
    post:
      operationId: appointmentsPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentCreate"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
          description: Prenotazione creata
      security:
      - bearerAuth: []
      summary: Crea prenotazione
      tags:
      - appointments
      x-eov-operation-handler: controllers/AppointmentsController
  /appointments/{appointmentId}:
    delete:
      operationId: appointmentsAppointmentIdDELETE
      parameters:
      - explode: false
        in: path
        name: appointmentId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Prenotazione cancellata
      security:
      - bearerAuth: []
      summary: Annulla prenotazione
      tags:
      - appointments
      x-eov-operation-handler: controllers/AppointmentsController
    put:
      operationId: appointmentsAppointmentIdPUT
      parameters:
      - explode: false
        in: path
        name: appointmentId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
        required: true
      responses:
        "200":
          description: Prenotazione aggiornata
      security:
      - bearerAuth: []
      summary: Reagenda prenotazione
      tags:
      - appointments
      x-eov-operation-handler: controllers/AppointmentsController
  /donors/{donorId}/tests:
    get:
      operationId: donorsDonorIdTestsGET
      parameters:
      - explode: false
        in: path
        name: donorId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/TestResult"
                type: array
          description: Lista risultati
      security:
      - bearerAuth: []
      summary: Elenco risultati esami del donatore (PDF/valori)
      tags:
      - tests
      x-eov-operation-handler: controllers/TestsController
  /notifications/subscribe:
    post:
      operationId: notificationsSubscribePOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_notifications_subscribe_post_request"
        required: true
      responses:
        "201":
          description: Token registrato
      security:
      - bearerAuth: []
      summary: Registra token per notifiche push
      tags:
      - notifications
      x-eov-operation-handler: controllers/NotificationsController
components:
  schemas:
    RegisterRequest:
      example:
        password: password
        codice_fiscale: codice_fiscale
        phone: phone
        dob: 2000-01-23
        last_name: last_name
        blood_group: A+
        gdpr_consent: true
        first_name: first_name
        email: email
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          format: email
          type: string
        phone:
          type: string
        codice_fiscale:
          description: Codice Fiscale italiano (sostituisce 'CPF')
          type: string
        dob:
          format: date
          type: string
        blood_group:
          example: A+
          type: string
        password:
          type: string
        gdpr_consent:
          type: boolean
      required:
      - codice_fiscale
      - email
      - first_name
      - gdpr_consent
      - last_name
      - password
      type: object
    LoginRequest:
      example:
        password: password
        email: email
      properties:
        email:
          format: email
          type: string
        password:
          type: string
      required:
      - email
      - password
      type: object
    Donor:
      example:
        codice_fiscale: codice_fiscale
        phone: phone
        dob: 2000-01-23
        last_name: last_name
        blood_group: blood_group
        gdpr_consent: true
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        donation_history:
        - date: 2000-01-23
          center_id: center_id
          type: whole_blood
          donation_id: donation_id
        - date: 2000-01-23
          center_id: center_id
          type: whole_blood
          donation_id: donation_id
        first_name: first_name
        email: email
      properties:
        id:
          format: uuid
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          format: email
          type: string
        phone:
          type: string
        codice_fiscale:
          type: string
        dob:
          format: date
          type: string
        blood_group:
          type: string
        donation_history:
          items:
            $ref: "#/components/schemas/DonationRecord"
          type: array
        gdpr_consent:
          type: boolean
      type: object
    DonorUpdate:
      example:
        phone: phone
        gdpr_consent: true
        email: email
      properties:
        phone:
          type: string
        email:
          type: string
        gdpr_consent:
          type: boolean
      type: object
    DonationRecord:
      example:
        date: 2000-01-23
        center_id: center_id
        type: whole_blood
        donation_id: donation_id
      properties:
        donation_id:
          type: string
        date:
          format: date
          type: string
        type:
          example: whole_blood
          type: string
        center_id:
          type: string
      type: object
    Center:
      example:
        address: address
        donation_types:
        - "[\"whole_blood\",\"plasma\"]"
        - "[\"whole_blood\",\"plasma\"]"
        lng: 6.027456183070403
        phone: phone
        name: name
        opening_hours: opening_hours
        id: id
        lat: 0.8008281904610115
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        lat:
          format: double
          type: number
        lng:
          format: double
          type: number
        phone:
          type: string
        opening_hours:
          type: string
        donation_types:
          items:
            example: "[\"whole_blood\",\"plasma\"]"
            type: string
          type: array
      type: object
    Slot:
      example:
        slot_id: slot_id
        start: 2000-01-23T04:56:07.000+00:00
        available: 0
        end: 2000-01-23T04:56:07.000+00:00
      properties:
        slot_id:
          type: string
        start:
          format: date-time
          type: string
        end:
          format: date-time
          type: string
        available:
          type: integer
      type: object
    AppointmentCreate:
      example:
        notes: notes
        center_id: center_id
        slot_id: slot_id
        donor_id: donor_id
      properties:
        donor_id:
          type: string
        center_id:
          type: string
        slot_id:
          type: string
        notes:
          type: string
      required:
      - center_id
      - donor_id
      - slot_id
      type: object
    Appointment:
      example:
        center_id: center_id
        appointment_id: appointment_id
        donor_id: donor_id
        slot:
          slot_id: slot_id
          start: 2000-01-23T04:56:07.000+00:00
          available: 0
          end: 2000-01-23T04:56:07.000+00:00
        status: booked
      properties:
        appointment_id:
          type: string
        donor_id:
          type: string
        center_id:
          type: string
        slot:
          $ref: "#/components/schemas/Slot"
        status:
          enum:
          - booked
          - cancelled
          - completed
          type: string
      type: object
    AppointmentUpdate:
      example:
        slot_id: slot_id
      properties:
        slot_id:
          type: string
      type: object
    TestResult:
      example:
        date: 2000-01-23
        summary: summary
        file_url: https://openapi-generator.tech
        test_id: test_id
      properties:
        test_id:
          type: string
        date:
          format: date
          type: string
        summary:
          type: string
        file_url:
          format: uri
          type: string
      type: object
    _auth_login_post_200_response:
      example:
        access_token: access_token
        token_type: Bearer
      properties:
        access_token:
          type: string
        token_type:
          example: Bearer
          type: string
      type: object
    _notifications_subscribe_post_request:
      properties:
        device_token:
          type: string
        platform:
          enum:
          - ios
          - android
          type: string
        preferences:
          type: object
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    oauth2:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.avis.it/authorize
          scopes:
            openid: OpenID Connect scope
            profile: Accesso profilo base
          tokenUrl: https://auth.avis.it/token
      type: oauth2
x-notes:
- branding: "Usare colori istituzionali AVIS: letters Pantone 300 (blu) + goccia Pantone\
    \ 032 (rosso). Vedere il manuale immagine coordinata AVIS per dettagli: https://www.avis.it/application/files/2016/0621/9615/Manuale_immagine_coordinata_2020.pdf"
- language: "Primary language: it-IT — All user-facing strings must be localizable\
    \ (it/en)."
- gdpr: Assicurarsi che tutti i consensi siano tracciati e che i dati sensibili (risultati
    esami) siano cifrati at-rest e in-transit.